backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "kagglehub>=0.3.13",
    "python-multipart>=0.0.20",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "sambanova>=1.2.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "weave>=0.52.17",
]
